Strip Path to Basename
To avoid writing the absolute path of a file to the PDS, strip things down to the base filename when we write. Also, to ensure we don't accidentally have a full path stored in the system that can overwrite a file unintentionally (i.e. someone used `write` manually rather than `writefile`), strip the stored filename to its base as well.

If an end user does wish to use an absolute (or relative) path when reading a file back out, they can still specify the path using the `--dest` flag.